backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
Pillow==10.4.0
httpx==0.27.0
authlib==1.3.1
itsdangerous==2.2.0
starlette==0.37.2
pydantic[email]==2.7.4
